Switch from uint32_t to double for math and return values

+++ nuttx/configs/stm32f777zit6-meadow/src/hcom_nx/hcom_nx_common.h
@@ -245,9 +245,9 @@ uint32_t meadow_parse_iso8601_time_period(const char *isoTimePeriod,
 
 // Meadow ADC
 int meadow_adc_configure(uint8_t gpioList[], uint32_t gpioCount,
-          uint16_t *userDataBuf, uint32_t adcBufSzBytes);
+          double *userDataBuf, uint32_t adcBufSzBytes);
 int meadow_adc_read_conversions(void);
-int meadow_adc_read_temp_vbat(uint32_t *batteryVoltage, uint32_t *temperatureValue);
+int meadow_adc_read_temp_vbat(double *batteryVoltage, double *temperatureValue);
 
 // Diagnostic related 
 int hcom_nx_diagnostic_app_execute(const HcomProtoHdrMsg_t *hdrMsg,
